Allow to make parsing hook based on PartyIdentification/ID
Some industries still identify party with a GLN and schemeID="GLN"
or with a VAT number and schemeID="<country code>:VAT". This allow to
process it with a parsing hook in the partner matching.

Add module partner_identification_import to match partner by external
identifier using the above hook.

fixup! Allow to make parsing hook based on PartyIdentification/ID